Crate xcm_docs

XCM Docs

Documentation and guides for XCM

Welcome to the Cross-Consensus Messaging documentation!

XCM is a language for communicating intentions between consensus systems. Whether you’re a developer, a blockchain enthusiast, or just interested in Polkadot, this guide aims to provide you with an easy-to-understand and comprehensive introduction to XCM.

Getting started

Head over to the fundamentals section. Then, go to the guides, to learn about how to do things with XCM.

Cookbook

There’s also the cookbook for useful recipes for XCM.

Glossary

There’s a glossary with common terms used throughout the docs.

Contribute

To contribute to the format, check out the RFC process. To contribute to these docs, make a PR.

Why Rust Docs?

Rust Docs allow docs to be as close to the source as possible. They’re also available offline automatically for anyone who has the polkadot-sdk repo locally.

Docs structure

flowchart
	docs[xcm_docs] --> fundamentals
	docs --> guides
	docs --> cookbook

Modules

cookbook

Useful recipes for programs and configurations.

fundamentals

Fundamentals of the XCM language. The virtual machine, instructions, locations and assets.

glossary

Glossary

guides

Step-by-step guides to set up an XCM environment and start hacking.